+++ src/app/welcome/welcome.component.ts
@@ -1,16 +1,19 @@
-import { Component, OnInit, NgZone } from '@angular/core';
+import { Component, OnInit, NgZone, ViewChild, ElementRef } from '@angular/core';
 import { RouterExtensions } from 'nativescript-angular/router';
 
 import { openUrl } from 'tns-core-modules/utils/utils';
 import { Page } from 'tns-core-modules/ui/page';
 import { TextField } from 'tns-core-modules/ui/text-field';
 import * as appSettings from 'tns-core-modules/application-settings';
+import { Color } from 'tns-core-modules/color';
 import { File, Folder } from 'tns-core-modules/file-system';
 
 import { BEANCOUNT_PATH_SETTING } from '../shared/constants';
 import { BeancountFileService } from '../shared/beancount-file.service';
 import { openFilePicker } from '../shared/beancount-file-picker';
 import { SideDrawerService } from '../shared/sidedrawer.service';
+
+const BEANCOUNT_WEBSITE = 'http://furius.ca/beancount/';
 
 @Component({
     selector: 'bc-welcome',
@@ -19,7 +22,13 @@ import { SideDrawerService } from '../shared/sidedrawer.service';
 })
 export class WelcomeComponent implements OnInit {
 
+    welcomeText = 'Beancount is a plain-text accounting system. ' +
+        'You can read more about it at the ' +
+        `<a href="${BEANCOUNT_WEBSITE}">Beancount website</a>.`;
     filePath: string;
+
+    @ViewChild('welcomeTextView', {static: false})
+    welcomeTextView: ElementRef;
 
     constructor(
         private ngZone: NgZone,
@@ -36,8 +45,19 @@ export class WelcomeComponent implements OnInit {
         this.filePath = this.beancountFile.path;
     }
 
+    onWelcomeTextLoaded() {
+        // https://github.com/NativeScript/NativeScript/issues/4746#issuecomment-475841566
+        const element = this.welcomeTextView.nativeElement.android;
+        // See welcome-text class in component CSS
+        element.setTextSize(16);
+        element.setLetterSpacing(0.04);
+        const textColor = new Color(255 * 0.6, 255, 255, 255).android;
+        element.setTextColor(textColor);
+        element.setLinkTextColor(textColor);
+    }
+
     openBeancountWebsite() {
-        openUrl('http://furius.ca/beancount/');
+        openUrl(BEANCOUNT_WEBSITE);
     }
 
     openFilePicker() {